+++ Bybit.Api.Examples/Usings.cs
@@ -4,6 +4,7 @@ global using Bybit.Api.Affiliate;
 global using Bybit.Api.Asset;
 global using Bybit.Api.CryptoLoan;
 global using Bybit.Api.Enums;
+global using Bybit.Api.InstitutionalLoan;
 global using Bybit.Api.Margin;
 global using Bybit.Api.Market;
 global using Bybit.Api.Position;