+++ Bybit.Api/Enums/BybitLoanRepayOrderStatus.cs
@@ -1,0 +1,13 @@
+namespace Bybit.Api.Enums;
+
+/// <summary>
+/// Institutional loan repayment request status.
+/// </summary>
+public enum BybitLoanRepayOrderStatus
+{
+    /// <summary>
+    /// Processing.
+    /// </summary>
+    [Map("P")]
+    Processing,
+}