+++ Bybit.Api/InstitutionalLoan/Requests/BybitLoanOrdersRequest.cs
@@ -1,0 +1,31 @@
+namespace Bybit.Api.InstitutionalLoan;
+
+/// <summary>
+/// Get institutional loan orders request.
+/// </summary>
+public record BybitLoanOrdersRequest
+{
+    /// <summary>
+    /// Loan order ID.
+    /// </summary>
+    [JsonProperty("orderId", NullValueHandling = NullValueHandling.Ignore)]
+    public string? OrderId { get; set; }
+
+    /// <summary>
+    /// Start timestamp in milliseconds.
+    /// </summary>
+    [JsonProperty("startTime", NullValueHandling = NullValueHandling.Ignore)]
+    public long? StartTime { get; set; }
+
+    /// <summary>
+    /// End timestamp in milliseconds.
+    /// </summary>
+    [JsonProperty("endTime", NullValueHandling = NullValueHandling.Ignore)]
+    public long? EndTime { get; set; }
+
+    /// <summary>
+    /// Limit for data size.
+    /// </summary>
+    [JsonProperty("limit", NullValueHandling = NullValueHandling.Ignore)]
+    public int? Limit { get; set; }
+}