+++ Bybit.Api/InstitutionalLoan/Requests/BybitLoanProductRequest.cs
@@ -1,0 +1,13 @@
+namespace Bybit.Api.InstitutionalLoan;
+
+/// <summary>
+/// Get institutional loan product information request.
+/// </summary>
+public record BybitLoanProductRequest
+{
+    /// <summary>
+    /// Product ID.
+    /// </summary>
+    [JsonProperty("productId", NullValueHandling = NullValueHandling.Ignore)]
+    public string? ProductId { get; set; }
+}