+++ Bybit.Api/InstitutionalLoan/Requests/BybitLoanRepayOrdersRequest.cs
@@ -1,0 +1,25 @@
+namespace Bybit.Api.InstitutionalLoan;
+
+/// <summary>
+/// Get institutional loan repayment orders request.
+/// </summary>
+public record BybitLoanRepayOrdersRequest
+{
+    /// <summary>
+    /// Start timestamp in milliseconds.
+    /// </summary>
+    [JsonProperty("startTime", NullValueHandling = NullValueHandling.Ignore)]
+    public long? StartTime { get; set; }
+
+    /// <summary>
+    /// End timestamp in milliseconds.
+    /// </summary>
+    [JsonProperty("endTime", NullValueHandling = NullValueHandling.Ignore)]
+    public long? EndTime { get; set; }
+
+    /// <summary>
+    /// Limit for data size.
+    /// </summary>
+    [JsonProperty("limit", NullValueHandling = NullValueHandling.Ignore)]
+    public int? Limit { get; set; }
+}